// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/BulletinAdminDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BulletinAdminDto = {
    bulletinId: string;
    title?: string | null;
    text?: string | null;
    active: boolean;
    startDate?: string | null;
    endDate?: string | null;
    createDate: string;
    modified: string;
    modifiedByUsername?: string | null;
};